chore: Moved date formatters to common folder
